pip_requirements.txt
faker==24.1.0
mysql-connector-python==8.2.0

